Resolve CSV Asan "use after free" memory issues
The recent changes to CSV reads were causing issues with ASAN testing
where we were getting a "heap use after free" error message in both
lscsv and get_info.

The offending code pattern was similar in both cases, even though the
pattern itself looks pretty innocent.

This fixes the ASAN issues by rewriting the code to be a little more
verbose and avoids the offending pattern.

When I have more time, I will try to see if we can reproduce this ASAN
issue outside of Arkouda (i.e. if it's a problem on the Chapel side)

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>