Closes #3425: Improve Msg Function Registration for Module Tracking (#3424)
* Only register Msg functions with a module name
With the `saveUsedModules` flag, we only want to be tracking
modules that can actually be included/excluded to avoid duplicating
function headers for those functions that are always included by
default.

By explicitly checking if the module name includes "Msg", we can
skip this for cases like `GenSymIO` which are always included and
was causing build issues with the module generation script.

* Add registration without module name