Closes #3378: Update chpl tutorial
This PR (closes #3378) is a rather substational update of the chpl tutorial since it's been quite a while. I added sections, changed explanations, and attempted to better motivate concepts instead of just introducing. I think this is a big improvement from what we had, but I did lose steam towards the end so I won't be surprised if there are lots of mistakes.

Given there's so many changes and moving things around, I think it probably easier to just look at the md file on my fork than the diff