Add initial implementation for Sparse Matrix Mult (#3516)
* Add initial implementation for Sparse Matrix Mult

Details in PR

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

* Address review feedback

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

* Fixing mypy and flake8 errors

mypy really wants the sparray type to be indexable, but that is not in
the scope of this PR. I can make it the case in a future PR, but for now
I would like to silence the error and move on.
Therefore I added a __get_item__ method to the sparray class which just
raises a NotImplementedError.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

* Resolve doc issues

This worked around the bug in chpldoc by not use'ing the layout enum

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

---------

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>