Closes #3049: upin `mypy` version (#3635)
This PR (closes #3049) removes the upper bound of our `mypy` version since the latest works just fine. Also related to #3346, I changed the python restriction to be `!=3.12.4` since `3.12.3` works just fine. It seems version `3.12.5` was just released, so we'll have to see if the forward refs are still an issue

Co-authored-by: Tess Hayes <stress-tess@users.noreply.github.com>