Fixes #3650: incorrect parquet reads with larger sizes
This PR fixes #3650 which caused incorrect results when using parquet to read a large array.

The issue was caused by moving the `readColumn` logic, which relies on `i` and updates `i`, into a function. The updates made to `i` passed into the function never made their way to `i` from the outer scope. This function is called inside a for loop over `num_row_groups`, so if there's more than one row group, the value of `i` gets reset. I resolved this by having the helpers return the updated value of `i`