Closes #3679: Flatten and unflatten runtime checks failure (#3680)
when building with with gasnet, runtime checks, and multi-dim, I was seeing some test failures with the following error:

```
arkouda//src/AryUtil.chpl:966: error: references to remote data cannot be passed to external routines like 'c_pointer_return'
```

This was caused by some `c_ptrTo` called on some remote values. This was definitely what was intended because we were passing the addresses to `get/put`. I looked into how this was handled in the aggregation code and found `getAddr` in `CommPrimitives`. I think this does the same thing as `c_ptrTo` but avoids the runtime check? I'm not sure but making the substitution seems to have fixed the issue

Co-authored-by: Tess Hayes <stress-tess@users.noreply.github.com>