Roll back test change to determine impact on testing (#3812)
A running mode check was added to correctly test with Client mode,
which allows developers to view server output when running the
test harness, but there have been some nightly testing failures
running in Server mode that are difficult to track down, due to
the lack of server output. Since this change seems possibly related,
we are temporarily reverting it to assess if this was causing those
failures.