Skip values and count values read properly
The variable skipIdx contains the number of values to be skipped in the column
prior to reading values. Skipping is done one row group at a time, so this
value must be updated as each row group is skipped.

Also, readColumnDbFl and readColumnIrregularBitWidth now return the number of
values read, so that ReadColumn increments the index into the output array
properly.