Read multiple row groups in Parquet files correctly (#3950)
* Read multiple row groups correctly

Iterate through the column's row groups while maintaining a count of the
total items read, and terminate the loop when the specified number of items
have been read.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>

* Skip values and count values read properly

The variable skipIdx contains the number of values to be skipped in the column
prior to reading values. Skipping is done one row group at a time, so this
value must be updated as each row group is skipped.

Also, readColumnDbFl and readColumnIrregularBitWidth now return the number of
values read, so that ReadColumn increments the index into the output array
properly.

---------

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>
Co-authored-by: John H. Hartman <jhh67@users.noreply.github.com>
Co-authored-by: ajpotts <amanda.j.potts@gmail.com>