Avoid 0 in test with unsigned numbers (#4104)
Adjusts a test to avoid accidentally trying to convert a -1 to an
unsigned number.

This test creates an array with random numbers between 0 and 100. But if
a number is 0, then the result after doing `0**4-1` is a negative
number, which causes exceptions when trying to cast the value to an
unsigned int.

Signed-off-by: Jade Abraham <jade.abraham@hpe.com>