Remove pinned Python versions and dependencies (#4097)
Removes some pinned Python versions and dependencies, allowing support
for python 3.13. This also removed `typed-ast`, which is not used
directly by arkouda and is EOL. In Python 3.8+, its recommend to use
`ast` as the replacement if its needed.

This PR adjusts the CI to test Python 3.12. Python 3.13 cannot be tested
in the CI yet due to [dependency issues from
numpy](https://github.com/Bears-R-Us/arkouda/issues/4098). numpy <2 does
work with python 3.13, but a wheel is not provided so it must be built
from source on `pip install`, which the CI cannot handle.

Resolves https://github.com/Bears-R-Us/arkouda/issues/3346: technically
was already resolved because ForwardRef was removed in
https://github.com/Bears-R-Us/arkouda/pull/3747

Signed-off-by: Jade Abraham <jade.abraham@hpe.com>