Improve the server launch scripts to handle launches on PBS (#4088)
This PR improves the server launch scripts in order to be able to run on
PBS systems in nightly testing. It includes two changes

- `ARKOUDA_SERVER_LAUNCH_PREFIX` environment can be used to "inject"
things before `./arkouda_server`: this will allow us to use
`chpl_launchcmd` testing script and pass some arguments to it.
- Fix how `-nl` is passed: Space-separated things has to be passed as
seperate items in the list when constructing the command.
`chpl_launchcmd` is sensitive to how `nl` argument is used, and was
getting tripped up if that argument is not passed properly.

I manually tested this on a PBS system. Next week, I am hoping to wire
nightly testing configs and test through that before merging this.

Signed-off-by: Engin Kayraklioglu <e-kayrakli@users.noreply.github.com>