Add `xp.trapz` (#4127)
Resolves: https://github.com/Bears-R-Us/arkouda/issues/4129

Add `trapz` (or `trapezoid`) function in Arkouda using the functionality
from the array API.
   
In the future the entirity of this function may be implemented purely in
Chapel.

See
https://numpy.org/doc/1.26/reference/generated/numpy.trapz.html#numpy.trapz
for reference.

Technically, `trapz` is deprecated in numpy `2.0` onwards, but Arkouda
still uses numpy `1.26`. I've added both overloads.

Also added tests.

---------

Signed-off-by: Shreyas Khandekar <khandeka@hotlum-login.head.cm.hpcrb.rdlabs.ext.hpe.com>
Co-authored-by: Shreyas Khandekar <khandeka@hotlum-login.head.cm.hpcrb.rdlabs.ext.hpe.com>