Closes 4206:  add install-pytables to Makefile (#4208)
This PR adds the ability to install `pytables` from source with a
makefile command `make install-pytables`. It also adds tests the CI. In
the case of almalinux, the `make install-hdf5` is moved to coincide as
it is a dependency of `pytables` that cannot be installed with `dnf` on
the current container version.

Closes 4206:  add install-pytables to Makefile

Co-authored-by: Amanda Potts <ajpotts@users.noreply.github.com>