Closes #4194:  add ruff-format to Makefile (#4200)
Adds `make ruff-format` to the Makefile. This is to ensure consistency
across contributions of formatting changes. Ran `make ruff-format`
across the project. Also, added `ruff format --check -diff` checks to
the CI. I switched from `black` to `ruff` because it is much faster and
was easier to integrate the config file with the project.

Closes #4194:  add ruff-format to Makefile

Co-authored-by: Amanda Potts <ajpotts@users.noreply.github.com>