Closes #4002: Test examples in arkouda.series.py (#4222)
Checked examples, and some of them are a little different from precisely
what comes out, but I think that can be excused because it looks a lot
nicer in the API. Fixed type annotations and also did some changes to
some of the docstrings themselves for typos. Probably doesn't correct
everything in the docstrings but at least puts a dent in it.

Closes #4002: Test examples in arkouda.series.py