Closes #4203: Fix max_bits on multidimensional arrays (#4204)
For some reason, `toSymEntry` has an input `param dimensions = 1`, but
no call (that I've seen) ever changes that. Trying to rework it so that
it doesn't need that and will just automatically convert it to whatever
dimension is in ndim (an attribute of `GenSymEntry`) seems impossible (I
tried, which is maybe not saying much). It seems to need to know at
compile time what kind of `SymEntry` it is returning. Engin pointed out
that this is solved by `instantiateAndRegister` (or just
`registerCommand` if that's all that's necessary).

This can probably be applied to many other instances where a call to
`toSymEntry` exists (e.g. I saw some in casting). This may not fix all
of them, but this alone should fix most of them and some further
tweaking should be possible on the others.

Since fixing all the ones in `BigIntMsg.chpl`, this also fixes the issue
from before where converting multidimensional bigint arrays to and from
numpy caused errors without so much reshaping.

Closes #4203: Fix max_bits on multidimensional arrays