Fix build issues cased by new makeDistArray
The new implementation of makeDistArray exposed that we were using a non
distributed sort in SegStringSort.chpl and this replaces that with the
dynamicRadixSort.

That in term caused issues with the module initialzation order which
cased build failures. This factors out the dynamicTwoArrayRadixSort into
it's own small module to avoid this issue

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>