Revert part of "Fix build issues cased by new makeDistArray"
This reverts commit b9449fcd846840646550b1f99c7a435476f7a9a2.

This commit tried to resolve build issues caused by the new `makeDistArray` fixes, but they seem to have broken argsort.

This reverts things to a state where we get a compile time warning about using a non-distributed sort in SegStringSort.chpl.

This means that that sort has been using a local version instead of distributed.