Resolve server crashes in SegStringSort (#4327)
Resolves https://github.com/Bears-R-Us/arkouda/issues/4318
Resolves https://github.com/Bears-R-Us/arkouda/issues/4326

However, this started causing server crashes for argsort on strings.

This is because there were cases where we were actually relying on the
buggy behavior of makeDistArray not creating a distributed array. And
when this behavior was fixed, this started causing server crashes. We
found the offending behavior and patched it in a current best effort to
get nightly testing back to blue.

Overall this is a strict improvement from what we had before, and should
also have a huge boost in performance for argsort in string since now
we're using a distributed sort instead of a local sort for
twoPhaseStringSort

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>