Closes #4320:  add pydocstyle to the project (#4324)
This PR adds pydocstyle to the project and adds an associated check to
the CI. pydocstyle checks docstring compliance with PEP 257, the
official Python style guide for docstrings. It also checks many of the
errors found by pydocstyle. Some error codes were added to the ignore
field and tickets were created to fix them separately.

Closes #4320:  add pydocstyle to the project

Co-authored-by: Amanda Potts <ajpotts@users.noreply.github.com>