Closes #4363:  add pydoclint to the project (#4364)
This PR adds pydoclint to the project. It also corrects errors that
occur when a docstring has a return section for a function with no
return.

pydoclint:
This tool is a fast docstring linter that checks whether docstring
sections (arguments, returns, raises, etc.) match the function signature
or implementation. It supports NumPy, Google, and Sphinx docstring
styles and can detect style violations that darglint might miss.

Closes #4363:  add pydoclint to the project

Co-authored-by: Amanda Potts <ajpotts@users.noreply.github.com>