Closes #4244: Make an unordered set union of two Strings arrays function (#4245)
Previously, `set_categories` was using several groupbys to set
categories for a Categorical array to some new set of categories. This
was inefficient. I figured out a new way to tackle this - union the
categories together (as Strings arrays), get the index of the old
categories in the new ones, and then remap. I was told that a `union1d`
function already exists, but looking into it, it did not seem to perform
very well. I created my own version that ignores order. Here's how it
works:

1. Union together the data within each locale across the arrays we're
doing the union over
2. Take the hash of each string modulo numLocales to determine which
locale to send each string to
3. After every locale sends all of its strings to all the other locales
(including itself), look at what everyone else has sent me
4. Each locale creates a new set containing the strings that were sent
to it (there are now no duplicates of strings across all locales)
5. Join the data together into a new SegString object

This seems to be the fastest way to do this. Getting incredible
performance out of this and major asymptotic improvement over `union1d`.

Closes #4244: Make an unordered set union of two Strings arrays function