Fixes bug involving arange and kwargs -- Closes 4321 (#4379)
Closes #4321 

This addresses the bug Engin noted in Issue 4321. It no longer uses
kwargs, and it should behave the same way numpy does. That is, I've
tested it with:

pda = ak.arange(10)         # same as if you said ak.arange(0,10,1)
pda = ak.arange(1,10)     # same as if you said ak.arange(1,10,1)
pda = ak.arange(1,10,2)

and in each case, it treats the numbers the same way numpy does.

There is still some anomalous behavior in here. It only handles integers
(I don't know the history behind that), and removing the **kwarg aspect
caused mypy issues because xp.arange calls it with floats. For that
matter, xp.arange claims to return floats, but it doesn't. In fact,
test_flip and test_roll fail if it returns floats.

Anomalous behavior aside (and perhaps that's for another issue), I
believe this particular bug is fixed.

Co-authored-by: drculhane <drculhane@users.noreply.github.com>
Co-authored-by: ajpotts <amanda.j.potts@gmail.com>