Closes #4386 add docstr-coverage (#4387)
This PR adds make `docstr-coverage`, which runs the `docstr-coverage`
tool over the `arkouda` directory and computes the percentage of
docstring coverage. It also adds a check to the CI to ensure coverage
stays above a certain level.

`docstr-coverage`: This library measures docstring coverage in Python
code, identifying functions, classes, methods, and modules lacking
docstrings. It provides statistics on overall docstring coverage for
files and projects.


Closes #4386 add docstr-coverage

Co-authored-by: Amanda Potts <ajpotts@users.noreply.github.com>