Detect server exit while reading connection file (#4403)
The testing helper `start_arkouda_server()` will now raise a
`RuntimeError` if the `arkouda_server` process exits without creating
the "connection info" file. It will also log the command used to launch
the server and the server output in this case.

Previously `start_arkouda_server()` would wait indefinitely for the
connection file to be created.

---------

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>