Part of #2384: Auto checkpoint upon exceeding memory percentage or idle time (#4391)
Address part of #2384. Replaces #4378.

The server now automatically saves a checkpoint if memory usage exceeded
a threshold or if the server has been idle for a bit. This is disabled
by default, see below. Requires `CheckpointMsg` module. Arkouda still
builds without this module.

The following server options in `CheckpointMsg` module control the
behavior:

* `checkpointMemPct` : Perform checkpointing automatically when memory
usage exceeds this many percent of available memory. By default / if
<=0, auto-checkpointing is not triggered by memory usage.

* `checkpointMemPctDelay` : When memory exceeds the `checkpointMemPct`
threhsold, wait for this many seconds of idle time before checkpointing.
By default / if <=0, uses 5 seconds.

* `checkpointIdleTime` : Perform checkpointing automatically when the
server is idle for this many seconds. By default / if <=0,
auto-checkpointing is not triggered by idle time.

* `checkpointCheckInterval` : The auto-checkpoint implementation will
wake up every this many seconds to check if a checkpoint needs to be
saved due to memory usage or idle time. By default / if <=0, uses min(
checkpointMemPctDelay, checkpointIdleTime ).

* `checkpointInterval` : Automatic checkpointing due to memory use or
idle time will wait for at least this many seconds after any completed
checkpoint save or load operations, whether automatic or
client-initiated. By default / if <=0, uses 3600 (one hour). This avoids
overly-frequent auto-checkpointing. The server does not need not be idle
during this interval. Checkpoint operations requested by the client are
not subject to this delay.

If the server receives a message from the client while
auto-checkpointing is in progress, it locks until auto-checkpointing
completes.

This PR also introduces a framework for client testing with customer
server options and invokes it as a sibling `pytest` command in
`Makefile`. Such testing is configured using `pytest.opts.ini` and
`tests/optioned-server/conftest.py` . It is used to test the
functionality added in this PR.

While there: switch the order of error() and critical() in
`src/Logging.chpl` to match the order in the enum; an analogous change
in `src/Message.chpl`; open the file in read mode, not append, in
`io_util.delimited_file_to_dict()`; added `io_util.directory_exists()`.

Next features to implement:
* The previous checkpoint is saved until the new checkpoint has
successfully completed.
* The server informs the client that it is auto-saving a checkpoint.
* Tune the default thresholds, esp. `checkpointInterval` .
* Do not checkpoint until the first client connects to the server and
has some activity.

The module-level variable `lastCkptCompletion` in CheckpointMsg.chpl
should be specific to each `DefaultServerDaemon` instance rather than be
a global. This is OK for now that we have only a single instance.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>