Adds negative axis coverage to ak.repeat, ak.argmin, ak.argmax, Closes 4407 and 4421 (#4408)
Closes #4407 and #4421

Numpy functions allow for negative numbers in the axis argument (e.g. -1
means the last axis, -2 the next-to-last, etc). Arkouda has such
functionality server-side, but it's not used consistently. This PR is
part of a series that aims to correct that.

Update: also adds negative axis coverage for argmin and argmax. Renames
the new chapel function validateNegativeAxis to validateNegativeAxes,
since "Axes" already existed, and this is just a case of overloading it.

Co-authored-by: drculhane <drculhane@users.noreply.github.com>
Co-authored-by: ajpotts <amanda.j.potts@gmail.com>