Closes #4235: searchsorted to work with uint type (#4440)
I think `searchsorted` should work with uint types now. The problem was
that somehow, `v = ak.array(v, dtype) + shift` would turn `v` into an
array of `float64` if `v` was originally `uint64` and `shift` was 0. I
changed it to be `v = ak.array(ak.array(v, dtype) + shift, dtype)` (and
similarly where we do the same thing to `a`) and now tests pass. This
should probably be fixed in `binopvs` or wherever it is as well, I
imagine this is probably not intended behavior.

I also added a check that the data types match in `searchsorted`.

Closes #4235: searchsorted to work with uint type