Fewer 'throws' while logging (#4433)
Switch logging methods to be non-throwing, by removing their dependence
on the `format()`. This eliminates the need for `try!` when calling them
and reduces (slightly) the size of the generated code.

While there, clean up and factor our the implementation; remove logger
calls in CheckpointMsg that were added by mistake.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>