Closes #2998: diff function (#4418)
The `diff` function already existed thanks to `array_api`, just added it
in here. Ordinarily it would be in some kind of utility functions, but
since that's where it's suggested that `any` and `all` be as well, I put
this in the same file. Mostly ripped off the existing unit tests but
added some things. Seemed to be unhappy with bool, so I made a small
change to Chapel code.

Closes #2998: diff function