Closes #4459: Simplify and extend logic in binopvv (#4462)
This cleans up the logic in `binopvv` and `doBinOpvv`. Currently, only
some types are supported due to the limitations of
`@arkouda.instantiateAndRegister`, but once that is changed, the
existing logic should more or less support additional sizes of `uint`
and `int`. The code will likely require significant refactoring, but the
logic should basically stay the same, and it should be much cleaner to
do it.

Closes #4459: Simplify and extend logic in binopvv