Closes #3964:  Refactor benchmark handling of running mode. (#4471)
This is a clean-up item that comments out arguments that are not
currently being used int `benchmark_v2/conftest.py`. I decided to
comment out rather than delete because I think they could be useful in
the future. However, leaving them in only partially implemented will be
confusing to the user.

Closes #3964:  Refactor benchmark handling of running mode.

Co-authored-by: ajpotts <ajpotts@users.noreply.github.com>