Part of #2384: Preserve previous auto-checkpoint (#4549)
Add the ability for `ak.save_checkpoint()` to preserve a checkpoint that
is about to be overwritten, by passing to it `mode='preserve_previous'`.

Switch automatic checkpointing to this mode by default instead of
`overwrite`. Add the server config `checkpointMode` to control this
setting.

While there, touch up ServerDaemon.chpl.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>