Closes #4468 aligning vecdot function to numpy (#4472)
Closes #4468 

-----

Update to the update

I rebased, and seem to have included changes that were already in the
most recent master.

-----

Update

The array_api tests had a test of vecdot which included the comment
"because numpy doesn't have a vecdot," and so it compared to something
that performed the old computation. That comment has been removed, and
the test now compares the result to np.vecdot.

-----

Still a draft, for one more round of tests.

The old vecdot has been removed. The new vecdot matches the performance
of numpy vecdot, and includes an optional axis parameter which can be
used when the input arrays have matching shape. This didn't require any
chapel-side code, just the use of broadcasting, binops, and ak.sum along
the last axis.

I've added a comment in the Notes section of the docstring describing
how to replicate the previous vecdot computation.

---------

Co-authored-by: 1RyanK <146032320+1RyanK@users.noreply.github.com>
Co-authored-by: drculhane <drculhane@users.noreply.github.com>