Closes 4520, incorporates where parameter in several functions where it had been missing (#4522)
Closes #4520 

This adds the where parameter to round, trunc, ceil and floor.

In src/EfuncMsg.chpl, those functions now refer to their input as "x"
instead of "pda." This allows us to fold all of them into the helper
function that was called _trig_helper, and is now called
_trig_hyp_round_helper. It has all the needed code to handle the where
parameter.

In arkouda/numpy/numeric.py, those functions now include a where
parameter and a descriptive docstring, and call _trig_hyp_round_helper
to interface to the server. Their input is called "x" instead of "pda,"
since that's what the helper function uses.

---------

Co-authored-by: drculhane <drculhane@users.noreply.github.com>