Closes #4497: Create a repartitionByLocaleString function (#4557)
Moved some of the logic from my `concatenateUniqueStrMsg` function into
the newly created `Repartition` module. No new unit tests because this
is still tested by `test_concatenate_uniquely`.

There are three coforalls in the `Repartition` version, but I think
ultimately, it is a small price to pay for the convenience of a helper
function in this context. Moreover, I don't think any of the coforalls
are particularly cumbersome.

I also revamped a `for` into two `forall`s, which if there are at least
two threads, I think that's a win.

Closes #4497: Create a repartitionByLocaleString function