Fix testing of auto checkpointing (#4592)
This PR fixes the shortcoming of this idiom in
`tests/optioned-server/auto-checkpoints.py`:

```python
assert directory_exists_delayed(autockptDir, 10)
delete_directory(autockptDir)
```

where the second line is executed after auto-checkpointing has started
and therefore the directory has been created, and before
auto-checkpointing is completed.

In such a case, removing the directory causes auto-checkpointing to
fail. This goes unnoticed because the client has no way to know about
it, however it throws off the carefully-carved sequence of events that
the test code is intended to cause / observe, and testing fails.

How to detect when the server has finished creating a checkpoint, which
we need before proceeding with next steps in testing? I added:
 `ak.client.wait_for_async_activity()`
that makes the server wait for an in-progress auto-checkpointing, if
any, to complete.

An alternative is for the server to create a marker file in the
checkpoint directory upon the completion of a checkpoint. I did not
implement this, however we may still want to add this. Another way is to
"tail" `arkouda.log` to see if it ends with a line containing "finished
autoCheckpoint".

Incidental changes:
* Add `currentTime()` as a proposed replacement for
`Time.timeSinceEpoch().totalSeconds()`.
* Prevent interruption of the server's "idle time" by a couple of
messages like `ruok` and `wait_for_async_activity`.
* Skip the processing of a message on the server if there is an error
early in message handling.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>