Closes #4623: Errors in multidimensional take testing (#4636)
Engin pointed out that multidim, multilocale tests were failing on
`test_take_multidim`. I think I found the cause of the issue.
Previously, `take` required that the indices had to have the same length
as the array it was being used on in the given axis. Since numpy has no
such restriction, I tossed that out and assumed it would be fine. But it
wasn't fine, because the way the source aggregator was being used, it
could be the case that source and dest could both be remote to the
locale doing the aggregation. This restructures the code so that it
should be okay. It passes tests with 16 locales on 3 dimensions.

Currently, there are a few places where the code could be more
efficient, but making it faster could be more difficult. So, this is
something of a quick fix just to get it working again.

Closes #4623: Errors in multidimensional take testing