Closes-2915, adds isnumeric function (#4694)
Update

--

The current revision (as of 7/10/25) moves isnumeric into the char
folder. It can now be addressed in three ways:

pda.isnumeric(), that is, as a method of the pdarray class
ak.char.isnumeric(pda), similar to np.char.isnumeric, and
ak.isnumeric(pda), and I'm not sure why that works.

arkouda/numpy/char.py has only the isnumeric function. Its input is
String, but when I import strings at the top of the file in order to use
typechecking, I get a circular import. The only way I found to avoid
that was not to use @typecheck, and instead to import strings inside the
function, and manually do the type check there.

--

I've implemented Ryan's suggestion to use a set instead of an array to
store the known numeric unicodes. I tried to search through all the
letters in a string with a forall, but I kept getting this error
message:

```
/home/drandy/projects/arkouda/src/SegmentedString.chpl:1776: error: A standalone or leader iterator is not found for the iterable expression in this forall loop
make: *** [Makefile:569: arkouda_server] Error 1

```
I understand that message, but it doesn't seem applicable, since it
refers to zipped iterations, which I don't think this is.

--

I've done a major rewrite of this, such that it now covers all the cases
that np.char.isnumeric() considers numeric (a total of 1922 characters,
including decimal digits, numerical superscripts and subscripts, special
characters with digits encircled or enclosed in parens, "vulgar
fractions," non-Arabic numerals, etc).

The function is now a string method (in strings.py, not numeric.py), so
it's addressed the same way the other string methods are, e.g.

stringName.isdigit()
stringName.isdecimal()
stringName.isnumeric()

It's also accessible as a standalone function, ak.isnumeric().

--

Closes #2915, adding an isnumeric function to match np.char.isnumeric.

Note:

This is presently in arkouda/numpy/numeric.py, but maybe it belongs in
strings.py? [_That's where it is now._]

There is no multi-dim test, because arkouda, AFAIK, doesn't support
multi-dim string pdarrays. String pdarrays are actually segmented
strings.

The SegString class already had a method isdigits, which I was able to
use to implement isnumeric. [_outdated. The class now has an isnumeric
method._]

---------

Co-authored-by: drculhane <drculhane@users.noreply.github.com>